+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.0] - 2023/04/19
+
+### Removed
+
+- Consulting, business, agency, and services pages.
+- Added a "503 under construction" page to replace the removed pages
+as a temporary handler.
+
 ## [1.0.4] - 2023/04/19
 
 ### Changes
@@ -67,7 +75,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Installed and configured `<gh-pages>`.
 - This CHANGELOG file to track the evolution of the project source code.
 
-[unreleased]: https://github.com/ogre2/Omega//compare/v1.0.4...HEAD
+[unreleased]: https://github.com/ogre2/Omega//compare/v1.1.0...HEAD
+[1.1.0]: https://github.com/ogre2/Omega/compare/v1.0.4...v1.1.0
 [1.0.4]: https://github.com/ogre2/Omega/compare/v1.0.3...v1.0.4
 [1.0.3]: https://github.com/ogre2/Omega/compare/v1.0.2...v1.0.3
 [1.0.2]: https://github.com/ogre2/Omega/compare/v1.0.1...v1.0.2